# a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: Discussly Web API
  description: Web Resources Specification (A7) for Discussly
servers:
  - url: 'http://lbaw2141-piu.lbaw-prod.fe.up.pt/'
    description: 'Production server'
...
paths:
...
  /register:
    get:
      operationId: R103
      summary: 'R103: Register Form'
      description: 'Provide new user registration form. Access: PUB'
      tags:
        - 'M01: Authentication and Individual Profile'
      responses:
        '200':
          description: 'Ok. Show [UI05](https://git.fe.up.pt/lbaw/lbaw2021/lbaw2141/-/raw/master/Wiki%20Images/signup_combined.png)'
        '400':
          description: 'Bad Request'
    post:
      operationId: R104
      summary: 'R104: Register Action'
      description: 'Processes the new user registration form submission. Access: PUB'
      tags:
        - 'M01: Authentication and Individual Profile'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                confirm_password:
                  type: string
                description:
                  type: string
                favourite_science:
                  type: boolean
                favourite_technology:
                  type: boolean
                favourite_engineering:
                  type: boolean
                favourite_maths:
                  type: boolean
              required:
                - username
                - email
                - password
                - confirm_password
                - favourite_science
                - favourite_techonolgy
                - favourite_engineering
                - favourite_maths
      responses:
        '302':
          description: Redirect after processing the new user information.
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: Successful authentication. Redirect to user profile.
                  value: '/users/{id}'
                302Failure:
                  description: Failed authentication. Redirect to login form.
                  value: /login
        '400':
          description: 'Bad Request'
...